# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ReSellution API
  version: 0.1.0
  description: Minimal OpenAPI spec for local development and API UI testing.
servers:
  - url: http://localhost:8080
    description: Local backend
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        email: { type: string, format: email }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        photo_url: { type: string, nullable: true }
      required: [id, full_name, email]
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: '#/components/schemas/User'
      required: [token, user]
    RegisterRequest:
      type: object
      properties:
        full_name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
      required: [full_name, email, password]
    LoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
      required: [email, password]
    UpdateProfileRequest:
      type: object
      properties:
        full_name: { type: string }
        city: { type: string }
        state: { type: string }
        bio: { type: string }
        photo_url: { type: string }
...
    ListingImage:
      type: object
      properties:
        id: { type: string, format: uuid }
        listing_id: { type: string, format: uuid }
        image_url: { type: string }
        position: { type: integer }
      required: [id, listing_id, image_url, position]
    Listing:
      type: object
      properties:
        id: { type: string, format: uuid }
        seller_id: { type: string, format: uuid }
        category_id: { type: string, format: uuid, nullable: true }
        title: { type: string }
        description: { type: string }
        condition: { type: string }
        price: { type: number }
        currency: { type: string }
        city: { type: string }
        state: { type: string, nullable: true }
        status:
          type: string
          enum: [active, reserved, sold, draft, deleted]
        view_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        images:
          type: array
          items:
            $ref: '#/components/schemas/ListingImage'
      required:
        - id
        - seller_id
        - title
        - description
        - condition
        - price
        - currency
        - city
        - status
        - created_at
        - updated_at
    CreateListingRequest:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        condition: { type: string }
        price: { type: number }
        currency: { type: string, default: INR }
        city: { type: string }
        state: { type: string }
        category_id: { type: string, format: uuid }
        status:
          type: string
          enum: [active, draft]
          default: active
      required: [title, description, condition, price, city]
    CreateListingResponse:
      type: object
      properties:
        listing:
          $ref: '#/components/schemas/Listing'
      required: [listing]
    ListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
        total: { type: integer }
      required: [listings, total]
    PaginatedListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        total_pages: { type: integer }
      required: [listings, total, page, limit, total_pages]
...
    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, reserved, sold, draft]
        sold_to_user_id:
          type: string
          format: uuid
      required: [status]
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
  /api/v1/users/me:
    patch:
      tags: [Users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
...
  /api/v1/listings/search:
    get:
      tags: [Listings]
      summary: Search listings
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: city
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Search result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingsResponse'
  /api/v1/listings:
    post:
      tags: [Listings]
      summary: Create listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateListingResponse'
        '401':
          description: Unauthorized
  /api/v1/listings/me:
    get:
      tags: [Listings]
      summary: List my listings
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [active, reserved, sold, draft, all]
      responses:
        '200':
          description: My listings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedListingsResponse'
        '401':
          description: Unauthorized
  /api/v1/listings/{id}/status:
    patch:
      tags: [Listings]
      summary: Update listing status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Listing updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  listing:
                    $ref: '#/components/schemas/Listing'
        '401':
          description: Unauthorized
  /api/v1/listings/{id}:
    delete:
      tags: [Listings]
      summary: Delete listing
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
  /api/v1/listings/{id}/images:
    post:
      tags: [Listings]
      summary: Upload listing image
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                position:
                  type: integer
      responses:
        '201':
          description: Image uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  image:
                    $ref: '#/components/schemas/ListingImage'
        '401':
          description: Unauthorized
  /api/v1/listings/{id}/report:
    post:
      tags: [Listings]
      summary: Report a listing
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingReportRequest'
      responses:
        '201':
          description: Listing reported
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '404':
          description: Listing not found
  /api/v1/admin/reports:
    get:
      tags: [AdminModeration]
      summary: List moderation reports
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [open, in_review, resolved, rejected, all]
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Reports page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsResponse'
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
  /api/v1/admin/reports/{id}:
    get:
      tags: [AdminModeration]
      summary: Get a moderation report
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Report detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    $ref: '#/components/schemas/Report'
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
        '404':
          description: Report not found
    patch:
      tags: [AdminModeration]
      summary: Update a report status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportStatusRequest'
      responses:
        '200':
          description: Report updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    $ref: '#/components/schemas/Report'
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
        '404':
          description: Report not found
  /api/v1/admin/reports/{id}/actions:
    post:
      tags: [AdminModeration]
      summary: Add a moderation action log entry
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModerationActionRequest'
      responses:
        '201':
          description: Action logged
          content:
            application/json:
              schema:
                type: object
                properties:
                  action:
                    $ref: '#/components/schemas/ModerationAction'
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
        '404':
          description: Report not found
  /api/v1/favorites:
    get:
      tags: [Favorites]
      summary: List favorites for the current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Favorites page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritesResponse'
        '401':
          description: Unauthorized
  /api/v1/favorites/{listing_id}:
    get:
      tags: [Favorites]
      summary: Check whether a listing is favorited by the current user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: listing_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Favorite status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteStatusResponse'
        '401':
          description: Unauthorized
    put:
      tags: [Favorites]
      summary: Favorite a listing
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: listing_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Favorited
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: Listing not found
    delete:
      tags: [Favorites]
      summary: Remove a listing from favorites
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: listing_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
  /api/v1/chat/conversations:
    get:
      tags: [Chat]
      summary: List conversations for the current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Conversation inbox
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationsResponse'
        '401':
          description: Unauthorized
    post:
      tags: [Chat]
      summary: Create or return a buyer conversation for a listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '200':
          description: Conversation created or already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation:
                    $ref: '#/components/schemas/ChatConversation'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Listing not found
  /api/v1/chat/conversations/{id}:
    get:
      tags: [Chat]
      summary: Get a single conversation with recent messages
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conversation detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation:
                    $ref: '#/components/schemas/ChatConversation'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Conversation not found
  /api/v1/chat/conversations/{id}/messages:
    get:
      tags: [Chat]
      summary: List messages for a conversation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Messages page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Conversation not found
    post:
      tags: [Chat]
      summary: Send a message in a conversation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    $ref: '#/components/schemas/ChatMessage'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Conversation not found
  /api/v1/chat/conversations/{id}/read:
    patch:
      tags: [Chat]
      summary: Mark all incoming messages in a conversation as read
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Marked read
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Conversation not found
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: List notifications for the current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: unread
          required: false
          schema:
            type: boolean
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Notifications page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
        '401':
          description: Unauthorized
  /api/v1/notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark a single notification as read
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification:
                    $ref: '#/components/schemas/Notification'
        '401':
          description: Unauthorized
        '404':
          description: Notification not found
  /api/v1/notifications/read-all:
    patch:
      tags: [Notifications]
      summary: Mark all notifications for the current user as read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notifications marked read
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
        '401':
          description: Unauthorized